requirements.txt
h5py==3.6.0
matplotlib==3.5.2
numpy==1.21.6
pandas==1.3.5
pesq==0.0.4
protobuf==3.20.1
pyroomacoustics==0.6.0
PySoundFile==0.9.0.post1
pystoi==0.3.3
pytorch-lightning==1.9.*
pydub
scipy==1.7.3
sdeint==0.2.4
SoundFile==0.10.3.post1
tensorboard==2.9.0
torch==1.11.0 --extra-index-url https://download.pytorch.org/whl/cu115
torch-ema==0.3
torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cu115
tqdm==4.63.0
wandb==0.13.10